+++ 404.php
@@ -4,7 +4,7 @@
 
     <?php AlisiosHooks::content_before(); ?>
 
-    <section id="primary" class="content col-xs-12 col-sm-8" role="main">
+    <section id="primary" <?php content_class(); ?> role="main">
 
         <?php AlisiosHooks::content_top(); ?>