+++ framework/alisios-functions-head.php
@@ -1,6 +1,36 @@
 <?php
 class AlisiosFunctionsHead {
 
+    /**
+     * Filter
+     * Creates a nicely formatted and more specific title element text
+     * for output in head of document, based on current view. Twenty Twelve 1.0
+     *
+     * @param string $title Default title text for current view.
+     * @param string $sep Optional separator.
+     * @return string Filtered title.
+     */
+    public static function wp_title( $title, $sep ) {
+        global $paged, $page;
+
+        if ( is_feed() )
+            return $title;
+
+        // Add the site name.
+        $title .= get_bloginfo( 'name' );
+
+        // Add the site description for the home/front page.
+        $site_description = get_bloginfo( 'description', 'display' );
+        if ( $site_description && ( is_home() || is_front_page() ) )
+            $title .= " $sep $site_description";
+
+        // Add a page number if necessary.
+        if ( $paged >= 2 || $page >= 2 )
+            $title .= " $sep " . sprintf( __( 'Page %s', ALISIOS_I18N ), max( $paged, $page ) );
+
+        return $title;
+    }
+
     /*
      * LESS
      * Load style.css or less.css, depend of var ALISIOS_USE_LESS
@@ -16,7 +46,7 @@ class AlisiosFunctionsHead {
     }
 
     /*
-     *
+     * Load JS and CSS libraries (JS on the bottom)
      */
     public static function enqueue_scripts_and_stylesheet() {
         //enqueue style