+++ framework/alisios-functions-header.php
@@ -14,3 +14,32 @@ class AlisiosFunctionsHeader {
         <?php
     }
 }
+
+/**
+ * Display the classes for the site element.
+ */
+function site_class( $class = '' ) {
+    // Separates classes with a single space, collates classes for body element
+    echo 'class="' . join(' ', get_site_class( $class )) . '"';
+}
+
+function get_site_class( $class = '' ) {
+    $classes = array();
+
+    $classes[] = 'site';
+    $classes[] = 'col-xs-12';
+    $classes[] = 'col-sm-12';
+
+    if(!empty($class)) {
+        if(!is_array($class))
+            $class = preg_split('#\s+#', $class);
+        $classes = array_merge($classes, $class);
+    } else {
+        // Ensure that we always coerce class to being an array.
+        $class = array();
+    }
+
+    $classes = array_map( 'esc_attr', $classes );
+
+    return apply_filters( 'site_class', $classes, $class );
+}