+++ framework/alisios-functions-init.php
@@ -26,6 +26,8 @@ define('ALISIOS_DIR_INC',       ALISIOS_DIR_FRAMEWORK . '/includes');
 require_once( ALISIOS_DIR_FRAMEWORK . '/alisios-functions-setup.php' );
 require_once( ALISIOS_DIR_FRAMEWORK . '/alisios-functions-head.php' );
 require_once( ALISIOS_DIR_FRAMEWORK . '/alisios-functions-header.php' );
+require_once( ALISIOS_DIR_FRAMEWORK . '/alisios-functions-content.php' );
+require_once( ALISIOS_DIR_FRAMEWORK . '/alisios-functions-sidebar.php' );
 require_once( ALISIOS_DIR_FRAMEWORK . '/alisios-functions-footer.php' );
 require_once( ALISIOS_DIR_FRAMEWORK . '/alisios-hooks.php' );